pyproject.toml
[tool.poetry]
name = "generator"
version = "0.1.0"
description = ""
authors = ["Kaj Siebert <kaj@k-si.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.38.0"
python-decouple = "^3.8"
pyyaml = "^6.0.1"
requests = "^2.32.3"
ratelimit = "^2.2.1"
pillow = "^10.4.0"
numpy = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
